+++ ReleaseNotes.md
@@ -3,3 +3,4 @@
 
 - [#305] Fix logging and error handling when a Deserialization error occurrs
 - [#304] Fix Deserialization error when FTP source is `ftp_estimated_source`
+- [#309] Fix `Found unknown distance unit m` and `kph`